refactor: rename MemberEntity
